+++ src/js/model/Lotto.js
@@ -12,7 +12,7 @@ export default class Lotto {
     return Math.floor(purchaseAmount / AMOUNT.UNIT);
   }
 
-  #generateLottoNumber() {
+  generateLottoNumber() {
     const { LENGTH_MIN, LENGTH_MAX, RANGE_MIN, RANGE_MAX } = LOTTO_NUMBER;
     const lottoNumbers = [...Array(RANGE_MAX)].map((_, index) => index + RANGE_MIN);
     shuffleArray(lottoNumbers);
@@ -20,6 +20,6 @@ export default class Lotto {
   }
 
   generateLottoTicket(count) {
-    this.#lottos = [...Array(count)].map(this.#generateLottoNumber);
+    this.#lottos = [...Array(count)].map(this.generateLottoNumber);
   }
 }
